+++ lib/hcaptcha_flutter.dart
@@ -1,1 +1,3 @@
 export 'src/hcaptcha_flutter.dart';
+export 'src/hcaptcha_flutter_web.dart'
+    if (dart.library.io) 'src/hcaptcha_flutter_method_channel.dart';