+++ lib/src/hcaptcha_flutter_platform_interface.dart
@@ -1,7 +1,10 @@
+import 'package:flutter/foundation.dart' show kIsWeb;
 import 'package:flutter/services.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
 
 import 'hcaptcha_flutter_method_channel.dart';
+import 'hcaptcha_flutter_web.dart'
+    if (dart.library.io) 'hcaptcha_flutter_method_channel.dart';
 
 typedef FlutterPluginHandler = Future<dynamic> Function(MethodCall call);
 
@@ -11,7 +14,8 @@ abstract class HCaptchaFlutterPlatform extends PlatformInterface {
 
   static final Object _token = Object();
 
-  static HCaptchaFlutterPlatform _instance = MethodChannelHCaptchaFlutter();
+  static HCaptchaFlutterPlatform _instance =
+      kIsWeb ? HCaptchaFlutterWeb() : MethodChannelHCaptchaFlutter();
 
   /// The default instance of [HCaptchaFlutterPlatform] to use.
   ///